requirements.txt
streamlit
mysql-connector-python
cryptography
Pillow
numpy
python-dotenv
